Big code clean up. Changes bundles extensions together, fixes .debug file, fixes background layer detection, replaces custom progressbar with native progressbar, removes useless code and files, uses global enviromental variables to store preview data
