+++ simpledissolve/SimpleDissolve/ApplyFilter/index.html
@@ -1,3 +1,5 @@
+<!doctype html>
+<!--
 /*************************************************************************
 * ADOBE CONFIDENTIAL
 * ___________________
@@ -11,28 +13,17 @@
 * then your use, modification, or distribution of it requires the prior
 * written permission of Adobe. 
 **************************************************************************/
+-->
+<html>
 
+<head>
+	<meta charset="utf-8">
+	<title></title>
+</head>
 
-#content {
-    margin-right:auto;
-    margin-left:auto;
-    vertical-align:middle;
-    width:100%;
-}
+<body>
+	<script src="js/libs/CSInterface.js"></script>
+	<script src="js/main.js"></script>
+</body>
 
-.debugtools{
-    background-color:#000;
-    position:absolute;
-    top:0px;
-    right:0px;
-    margin: 0;
-    padding: 0;
-    list-style-type: none;
-    text-align: center;
-}
-
-.debugtools li {
-    display:inline;
-    margin:3px 0px 3px 0px;
-}
-
+</html>
