+++ simpledissolve/SimpleDissolve/CSXS/manifest.xml
@@ -1,4 +1,5 @@
-<?xml version="1.0" encoding="UTF-8"?><!--
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
 /*************************************************************************
 * ADOBE CONFIDENTIAL
 * ___________________
@@ -13,14 +14,15 @@
 * written permission of Adobe. 
 **************************************************************************/
 -->
-<ExtensionManifest Version="5.0" ExtensionBundleId="SimpleDissolveUI" ExtensionBundleVersion="1.0.0"
-		ExtensionBundleName="SimpleDissolveUI" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
+<ExtensionManifest Version="5.0" ExtensionBundleId="com.adobe.SimpleDissolve" ExtensionBundleVersion="1.0.0"
+ ExtensionBundleName="SimpleDissolve" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
 	<ExtensionList>
-		<Extension Id="SimpleDissolveUI" Version="1.0" />
+		<Extension Id="com.adobe.SimpleDissolve.ApplyFilter" Version="1.0" />
+		<Extension Id="com.adobe.SimpleDissolve.UI" Version="1.0" />
 	</ExtensionList>
 	<ExecutionEnvironment>
 		<HostList>
-            <Host Name="PHXS" Version="[14.0,99.0]" />
+			<Host Name="PHXS" Version="[17.0,99.0]" />
 		</HostList>
 		<LocaleList>
 			<Locale Code="All" />
@@ -30,11 +32,31 @@
 		</RequiredRuntimeList>
 	</ExecutionEnvironment>
 	<DispatchInfoList>
-		<Extension Id="SimpleDissolveUI">
+		<Extension Id="com.adobe.SimpleDissolve.ApplyFilter">
+			<DispatchInfo>
+				<Resources>
+					<MainPath>./ApplyFilter/index.html</MainPath>
+					<ScriptPath>./ApplyFilter/jsx/hostscript.jsx</ScriptPath>
+				</Resources>
+				<Lifecycle>
+					<AutoVisible>false</AutoVisible>
+				</Lifecycle>
+				<UI>
+					<Type>Custom</Type>
+					<Geometry>
+						<Size>
+							<Height>1</Height>
+							<Width>1</Width>
+						</Size>
+					</Geometry>
+				</UI>
+			</DispatchInfo>
+		</Extension>
+		<Extension Id="com.adobe.SimpleDissolve.UI">
 			<DispatchInfo >
 				<Resources>
-				<MainPath>./index.html</MainPath>
-				<ScriptPath>./jsx/hostscript.jsx</ScriptPath>
+				<MainPath>./UI/index.html</MainPath>
+				<ScriptPath>./UI/jsx/hostscript.jsx</ScriptPath>
 				</Resources>
 				<Lifecycle>
 					<AutoVisible>true</AutoVisible>
