--- simpledissolve/SimpleDissolveApplyFilter/.debug
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?> 
-<ExtensionList>
-    <Extension Id="SimpleDissolveSupport">
-        <HostList>
-           
-            <!-- Comment Host tags according to the apps you want your panel to support -->
-            
-            <!-- Photoshop -->
-            <Host Name="PHXS" Port="12122"/>
-                        
-        </HostList>
-    </Extension>
-</ExtensionList>
-
