--- simpledissolve/SimpleDissolveApplyFilter/CSXS/manifest.xml
@@ -1,58 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-/*************************************************************************
-* ADOBE CONFIDENTIAL
-* ___________________
-*
-* Copyright 2014 Adobe
-* All Rights Reserved.
-*
-* NOTICE: Adobe permits you to use, modify, and distribute this file in
-* accordance with the terms of the Adobe license agreement accompanying
-* it. If you have received this file from a source other than Adobe,
-* then your use, modification, or distribution of it requires the prior
-* written permission of Adobe. 
-**************************************************************************/
--->
-<ExtensionManifest Version="5.0" ExtensionBundleId="SimpleDissolveSupport" ExtensionBundleVersion="1.0.0"
-		ExtensionBundleName="SimpleDissolveApplyFilter" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
-	<ExtensionList>
-		<Extension Id="SimpleDissolveApplyFilter" Version="1.0" />
-	</ExtensionList>
-	<ExecutionEnvironment>
-		<HostList>
-            <Host Name="PHXS" Version="[14.0,99.0]" />
-		</HostList>
-		<LocaleList>
-			<Locale Code="All" />
-		</LocaleList>
-		<RequiredRuntimeList>
-			<RequiredRuntime Name="CSXS" Version="5.0" />
-		</RequiredRuntimeList>
-	</ExecutionEnvironment>
-	<DispatchInfoList>
-		<Extension Id="SimpleDissolveApplyFilter">
-			<DispatchInfo >
-				<Resources>
-				<MainPath>./index.html</MainPath>
-				<ScriptPath>./jsx/hostscript.jsx</ScriptPath>
-				</Resources>
-				<Lifecycle>
-					<AutoVisible>false</AutoVisible>
-					<StartOn>
-						<Event>com.adobe.event.loadSimpleDissolveSupport</Event>
-					</StartOn>
-				</Lifecycle>
-				<UI>
-					<Type>Panel</Type>
-					<Geometry>
-						<Size>
-							<Height>1</Height>
-							<Width>1</Width>							
-						</Size>						
-					</Geometry>
-				</UI>
-			</DispatchInfo>
-		</Extension>
-	</DispatchInfoList>
-</ExtensionManifest>
