--- simpledissolve/SimpleDissolveApplyFilter/index.html
@@ -1,29 +1,0 @@
-<!doctype html>
-<!--
-/*************************************************************************
-* ADOBE CONFIDENTIAL
-* ___________________
-*
-* Copyright 2014 Adobe
-* All Rights Reserved.
-*
-* NOTICE: Adobe permits you to use, modify, and distribute this file in
-* accordance with the terms of the Adobe license agreement accompanying
-* it. If you have received this file from a source other than Adobe,
-* then your use, modification, or distribution of it requires the prior
-* written permission of Adobe. 
-**************************************************************************/
--->
-<html>
-<head>
-<meta charset="utf-8">
-<title></title>
-</head>
-
-<body>
-	<script src="js/libs/CSInterface.js"></script>
-	<script src="js/libs/jquery-2.0.2.min.js"></script>
-	<script src="js/main.js"></script>
-</body>
-</html>
-
