--- simpledissolve/SimpleDissolveUI/.debug
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?> 
-<ExtensionList>
-    <Extension Id="SimpleDissolve">
-        <HostList>
-            <Host Name="PHXS" Port="12121"/>       
-        </HostList>
-    </Extension>
-</ExtensionList>
-
