--- simpledissolve/SimpleDissolveUI/css/theme.css
@@ -1,108 +1,0 @@
-/*************************************************************************
-* ADOBE CONFIDENTIAL
-* ___________________
-*
-* Copyright 2014 Adobe
-* All Rights Reserved.
-*
-* NOTICE: Adobe permits you to use, modify, and distribute this file in
-* accordance with the terms of the Adobe license agreement accompanying
-* it. If you have received this file from a source other than Adobe,
-* then your use, modification, or distribution of it requires the prior
-* written permission of Adobe. 
-**************************************************************************/
-html, body {
-	float:left;
-	height: 100%;
-	margin: 0px;
-	overflow: hidden;
-	width: 100%;
-	font-size: 13px;
-	color: #000000;
-	-webkit-touch-callout: none;
-	-webkit-user-select: none;
-}
-
-.divMain {
-	height: 100%;
-	width: 100%;
-	overflow: hidden;
-	padding-top: 0px;
-}
-
-.divPreview {
-	float: left;
-	height: 100%;
-	width: 190px;
-	vertical-align: middle;
-}
-
-.divControls {
-	float: left;
-	margin: 0px;
-	height: 100%;
-	width: 150px;
-	margin-top: 10px;
-}
-
-.divBtnBar {
-	float:left;
-	height: 100%;
-	width: 105px;
-}
-
-#cnvsPreview {
-	margin-right: 10px;
-	margin-left: 5px;
-}
-
-#iDissolvePc {
-	width: 60px;
-	height: 20px;
-}
-
-button {
-	float: right;
-	border-radius: 2px;
-	line-height: 0px;
-	height: 22px;
-	width: 74px;
-	margin: 5px;
-	background: #f0f0f0;
-	border: 1px solid #7d7c7d;
-}
-
-button:active {
-	background: #9b9b9b;
-}
-
-.okBtn {
-	height: 23px;
-	width: 75px;
-	margin: 5px;
-	border: 2px solid #7d7c7d;
-}
-
-.olclass{
-	list-style: none;
-	padding: 0px;
-	margin-left: 0px;
-	margin-top: 5px;
-	margin-bottom: 0px;
-	width: 100%;
-}
-
-li {
-	padding-bottom: 10px;
-}
-
-#fsDisposition {
-	margin: 0px;
-	margin-top: 10px;
-}
-
-input[type=number]::-webkit-inner-spin-button, 
-input[type=number]::-webkit-outer-spin-button { 
-  -webkit-appearance: none; 
-  margin: 0; 
-}
